+++ zcode/inout/inout_core.py
@@ -29,6 +29,7 @@ from datetime import datetime
 import os
 import sys
 import logging
+import warnings
 import numpy as np
 
 __all__ = ['Keys', 'MPI_TAGS', 'StreamCapture', 'bytesString', 'getFileSize', 'countLines', 'estimateLines',
@@ -267,18 +268,18 @@ def dictToNPZ(dataDict, savefile, verbose=False, log=None):
     If the path to the given filename doesn't already exist, it is created.
     If ``verbose`` is True, the saved file size is printed out.
     """
-
     # Make sure path to file exists
     checkPath(savefile)
-
     # Make sure there are no scalars in the input dictionary
     for key, item in dataDict.items():
         if np.isscalar(item):
-            raise ValueError("Value '%s' for key '%s' is a scalar.  Not allowed." % (str(item), str(key)))
+            warnStr = "Value '%s' for key '%s' is a scalar." % (str(item), str(key))
+            warnings.warn(warnStr)
+            dataDict[key] = np.array(item)
 
     # Save and confirm
     np.savez(savefile, **dataDict)
-    if(not os.path.exists(savefile)):
+    if not os.path.exists(savefile):
         raise RuntimeError("Could not save to file '%s'." % (savefile))
 
     logStr = " - Saved dictionary to '%s'" % (savefile)
@@ -314,7 +315,7 @@ def npzToDict(npz):
     for key in list(npz.keys()):
         vals = npz[key]
         # Extract objects (e.g. dictionaries) packaged into size=1 arrays
-        if(np.size(vals) == 1 and (type(vals) == np.ndarray or type(vals) == np.array) and 
+        if(np.size(vals) == 1 and (type(vals) == np.ndarray or type(vals) == np.array) and
            vals.dtype.type is np.object_):
             vals = vals.item()